+++ src/actions/createUser.ts
@@ -1,19 +1,20 @@
 "use server";
 
 import { prisma } from "@/lib";
-import { registerFormValidationsSchema } from "@/utils";
+import { registerFormValidationsSchema } from "@/lib";
 import { redirect } from "next/navigation";
 import { hash } from "bcryptjs";
+import { RoutesPage } from "@/types";
 
 export async function createUser(formData: FormData) {
   const values = Object.fromEntries(formData.entries());
 
   const { username, email, password } =
-    registerFormValidationsSchema.cast(values);
+    registerFormValidationsSchema.parse(values);
 
   const hashedPassword = await hash(password, 12);
 
-  await prisma.user.create({
+  const result = await prisma.user.create({
     data: {
       username,
       email,
@@ -21,5 +22,7 @@ export async function createUser(formData: FormData) {
     },
   });
 
-  redirect("/profile");
+  if (result) {
+    redirect(RoutesPage.LOGIN);
+  }
 }