+++ src/components/Buttons/index.ts
@@ -1,1 +1,2 @@
-export * from "./AuthButtons";
+export * from "./FormSubmitButton";
+export { default as LoadingButton } from "./LoadingButton";