+++ src/components/NavBar/index.tsx
@@ -12,23 +12,23 @@ import {
 } from "@nextui-org/react";
 import { Brand } from "../Brand";
 import { ThemeSwitcher } from "../ThemeSwitcher";
-import { Search } from "../Search";
 import { RoutesPage } from "@/types";
 import { MENU_ITEMS } from "@/constants";
-import { usePathname } from "next/navigation";
-import clsx from "clsx";
+import { usePathname, useRouter } from "next/navigation";
 import { useSession, signOut } from "next-auth/react";
-import {
-  CircleUserRound,
-  ShoppingCart,
-  Search as SearchIcon,
-  LogOut,
-} from "lucide-react";
+import { CircleUserRound, ShoppingCart, LogOut } from "lucide-react";
+import { cn } from "@/utils";
 
 export const NavBar: React.FC = () => {
   const { data: session } = useSession();
+  const router = useRouter();
   const pathname = usePathname();
   const [isMenuOpen, setIsMenuOpen] = useState(false);
+
+  const handleSignOut = () => {
+    signOut();
+    router.push(RoutesPage.HOME);
+  };
 
   return (
     <Navbar
@@ -43,14 +43,14 @@ export const NavBar: React.FC = () => {
         <Brand />
       </NavbarBrand>
 
-      <NavbarContent
+      {/* <NavbarContent
         justify="start"
         className="hidden sm:flex"
       >
         <NavbarItem className="w-full">
           <Search />
         </NavbarItem>
-      </NavbarContent>
+      </NavbarContent> */}
       <NavbarContent
         justify="end"
         className="gap-4"
@@ -90,13 +90,13 @@ export const NavBar: React.FC = () => {
             <Link
               href={RoutesPage.HOME}
               color="secondary"
-              onClick={() => signOut()}
+              onClick={handleSignOut}
             >
               <LogOut color="#FFA500" />
             </Link>
           </NavbarItem>
         )}
-        <NavbarItem className="sm:hidden">
+        {/* <NavbarItem className="sm:hidden">
           <Link
             href="#"
             color="secondary"
@@ -107,7 +107,7 @@ export const NavBar: React.FC = () => {
               color="#FFA500"
             />
           </Link>
-        </NavbarItem>
+        </NavbarItem> */}
         <NavbarItem>
           <ThemeSwitcher />
         </NavbarItem>
@@ -118,9 +118,10 @@ export const NavBar: React.FC = () => {
             <NavbarMenuItem key={`${menuItem.label}-${index}`}>
               <Link
                 color="secondary"
-                className={clsx("w-full", {
-                  "text-blue-600": pathname === menuItem.href,
-                })}
+                className={cn(
+                  "w-full",
+                  pathname === menuItem.href && "text-blue-600"
+                )}
                 href={menuItem.href as string}
                 size="sm"
               >