+++ src/components/UI/index.ts
@@ -1,0 +1,4 @@
+export * from "./button";
+export { default as H1 } from "./h1";
+export * from "./label";
+export * from "./select";