+++ src/lib/auth.ts
@@ -4,12 +4,9 @@ import CredentialsProvider from "next-auth/providers/credentials";
 import type { NextAuthOptions } from "next-auth";
 
 export const authOptions: NextAuthOptions = {
-  session: {
-    strategy: "jwt",
-  },
   providers: [
     CredentialsProvider({
-      name: "Sign in",
+      name: "Credentials",
       credentials: {
         email: {
           label: "Email",
@@ -29,47 +26,46 @@ export const authOptions: NextAuthOptions = {
           },
         });
 
-        if (!userFound) throw new Error("Usuario no encontrado");
+        if (!userFound) throw new Error("Usuario no encontrado!");
 
         const matchPassword = await compare(
           credentials.password,
           userFound.password!
         );
 
-        if (!matchPassword) throw new Error("Contraseña incorrecta");
+        if (!matchPassword) throw new Error("Contraseña incorrecta!");
 
         return {
           id: userFound.id,
           email: userFound.email,
           username: userFound.username,
-          randomKey: "Hey cool",
         };
       },
     }),
   ],
-  callbacks: {
-    session: ({ session, token }) => {
-      return {
-        ...session,
-        user: {
-          ...session.user,
-          id: token.id,
-          randomKey: token.randomKey,
-        },
-      };
-    },
-    jwt: ({ token, user }) => {
-      if (user) {
-        const u = user as unknown as any;
-        return {
-          ...token,
-          id: u.id,
-          randomKey: u.randomKey,
-        };
-      }
-      return token;
-    },
-  },
+  // callbacks: {
+  //   session: ({ session, token }) => {
+  //     return {
+  //       ...session,
+  //       user: {
+  //         ...session.user,
+  //         id: token.id,
+  //         randomKey: token.randomKey,
+  //       },
+  //     };
+  //   },
+  //   jwt: ({ token, user }) => {
+  //     if (user) {
+  //       const u = user as unknown as any;
+  //       return {
+  //         ...token,
+  //         id: u.id,
+  //         randomKey: u.randomKey,
+  //       };
+  //     }
+  //     return token;
+  //   },
+  // },
   pages: {
     signIn: "/login",
   },