+++ src/lib/index.ts
@@ -1,2 +1,6 @@
 export * from "./auth";
+export * from "./cities";
+export * from "./filter-products.validation";
+export * from "./forms.validations";
 export * from "./prismaClient";
+export * from "./token-service";