+++ src/types/product-filter-values.type.ts
@@ -1,0 +1,4 @@
+import { productFilterSchema } from "@/lib";
+import { z } from "zod";
+
+export type ProductFilterValues = z.infer<typeof productFilterSchema>;