+++ src/types/routes-page.enum.ts
@@ -5,4 +5,8 @@ export enum RoutesPage {
   PROFILE = "/profile",
   FAQ = "/faq",
   ADMIN = "/admin",
+  FORGET_PASSWORD = "/forget-password",
+  REGISTER = "/register",
+  PRODUCTS = "/products",
+  UPDATE_PASSWORD = "/update-password",
 }