+++ CHANGELOG.md
@@ -2,6 +2,43 @@
 
 All notable changes to this project will be documented in this file. See [commit-and-tag-version](https://github.com/absolute-version/commit-and-tag-version) for commit guidelines.
 
+## 0.1.3 (2024-04-07)
+
+
+### Features
+
+* adding authentication ([c95c827](https://github.com/hugocruzlfc/tu-merca/commit/c95c827c6ce02fdbeb53d6c38a85831771ecdf34))
+* adding brand icon for current theme ([cb33846](https://github.com/hugocruzlfc/tu-merca/commit/cb3384641ecb0b6f583a89a1f35e7ecda350e819))
+* adding create user action ([0f323e3](https://github.com/hugocruzlfc/tu-merca/commit/0f323e317fc4e2eb9c6b07a83a8f77949be28f01))
+* adding filter behavior ([8e20d53](https://github.com/hugocruzlfc/tu-merca/commit/8e20d53888f3c8840ff8b180077f92e4d8450f47))
+* adding footer and icons ([1912849](https://github.com/hugocruzlfc/tu-merca/commit/1912849f729140ae38dc5fa34fcaa95517214f6b))
+* adding formik and yup libraries and sign main components ([a84091c](https://github.com/hugocruzlfc/tu-merca/commit/a84091cac1c48ff00f9f083ef365552db75c19c1))
+* adding had visited navbar link ([b31cada](https://github.com/hugocruzlfc/tu-merca/commit/b31cada9091193f2ce361948818250f562754b1c))
+* adding husky pre commit ([3b282ff](https://github.com/hugocruzlfc/tu-merca/commit/3b282ff6f8acbf20d6bbfcbda5f4a1721e1aad3e))
+* adding login page ([2407b2c](https://github.com/hugocruzlfc/tu-merca/commit/2407b2c8346c7d72bce44587126f069666d81195))
+* adding meta data default values ([3c2edd8](https://github.com/hugocruzlfc/tu-merca/commit/3c2edd80228ffc3b002e51a0210be1b45c1f451c))
+* adding meta icons ([ccd8568](https://github.com/hugocruzlfc/tu-merca/commit/ccd85689cb86159c548f5d25bbf58f8fed8fb895))
+* adding meta nextui configurations ([99d8b23](https://github.com/hugocruzlfc/tu-merca/commit/99d8b23aaa70492cd34cf903ce1a71d67c35a5ef))
+* adding metadata to remain pages ([c4080f1](https://github.com/hugocruzlfc/tu-merca/commit/c4080f102c68d4f1bea05ab2081c07e5bbc1e722))
+* adding migration and seed ([e49416e](https://github.com/hugocruzlfc/tu-merca/commit/e49416e34d7703151b1837b996544a43ad7c7e89))
+* adding nav bar ([b850f44](https://github.com/hugocruzlfc/tu-merca/commit/b850f44544d460ee11a27027f511df2391202e81))
+* adding primsa and vercel connect ([03966b6](https://github.com/hugocruzlfc/tu-merca/commit/03966b6860af905179a7b6b7f17833d85e199e3f))
+* adding prisma configuration and docker compose file for local ([ae75574](https://github.com/hugocruzlfc/tu-merca/commit/ae7557478abc47ddb7a4c2db1aca9c69da7cc32a))
+* adding products component ([de9afe6](https://github.com/hugocruzlfc/tu-merca/commit/de9afe6083b6862953aae153ab52af1016dc2740))
+* adding register and auth api ([fa9c13b](https://github.com/hugocruzlfc/tu-merca/commit/fa9c13bea8261f75daf4ebf6bfc86fedc885034f))
+* adding toast notifications ([f5e073d](https://github.com/hugocruzlfc/tu-merca/commit/f5e073d32af8b3ac1bf7e11750c493e711435d17))
+* products page and metadata ([9a560a8](https://github.com/hugocruzlfc/tu-merca/commit/9a560a8b065c3dbf3f6cb26488f7af511b190f68))
+* protected routes with middleware ([f185df2](https://github.com/hugocruzlfc/tu-merca/commit/f185df23113b9beebcd8a0eb92cfa520ffb2e77f))
+* update project and db ([e81249b](https://github.com/hugocruzlfc/tu-merca/commit/e81249b300e0bb0ed5557f5cb6472063ba416dae))
+
+
+### Bug Fixes
+
+* adding .env to gitignore ([1ad73f3](https://github.com/hugocruzlfc/tu-merca/commit/1ad73f341b9b5e2ef99e2a5ea4437a9bcdcf2c90))
+* adding handler environment ([243e42a](https://github.com/hugocruzlfc/tu-merca/commit/243e42aeffea29bbdcebc53aad811758ba86e7ef))
+* update middleware matcher ([62fadea](https://github.com/hugocruzlfc/tu-merca/commit/62fadeaa0dcb270300105e1aaa000f42fab14bac))
+* update migration init ([afb1543](https://github.com/hugocruzlfc/tu-merca/commit/afb1543a9c29fd141566a356a377e76398b584bf))
+
 ## 0.1.2 (2024-04-04)